Move SetupPanelModel to be closer to the class that uses it (MainPanel)
