+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/MetaSetupPanel.java
@@ -1,7 +1,7 @@
 package games.strategy.engine.framework.startup.ui;
 
 import games.strategy.engine.framework.startup.launcher.ILauncher;
-import games.strategy.engine.framework.startup.mc.SetupPanelModel;
+import games.strategy.engine.framework.startup.ui.panels.main.SetupPanelModel;
 import games.strategy.triplea.UrlConstants;
 import games.strategy.triplea.settings.ClientSetting;
 import java.awt.Font;
